refactor: extract LoopBrowserItems from LoopBrowser (647→402 lines)
New file: src/components/assets/LoopBrowserItems.tsx (257 lines)
- MiniWaveform, SvgMiniWaveform, fmtDuration
- PresetLoopItem, AssetLoopItem

LoopBrowser.tsx: 647 → 402 lines (38% reduction)
Build: 0 errors. Tests: 46/46 pass.
