+++ test-results/transport-Transport-Contro-558a3-ton-exists-and-is-clickable-chromium/error-context.md
@@ -1,0 +1,18 @@
+# Page snapshot
+
+```yaml
+- generic [ref=e3]:
+  - generic [ref=e4]: "[plugin:vite:react-babel] /Users/junmingong/.openclaw/workspace/acestep-daw/src/components/assets/LoopBrowserItems.tsx: 'import' and 'export' may only appear at the top level. (6:0) 9 |"
+  - generic [ref=e5]: /Users/junmingong/.openclaw/workspace/acestep-daw/src/components/assets/LoopBrowserItems.tsx:6:0
+  - generic [ref=e6]: "4 | */ 5 | export function MiniWaveform({ data, color, height = 32 }: { data: number[] | null; color: string; height?: number }) { 6 | import { useCallback, useEffect, useRef, useState } from 'react'; | ^ 7 | import type { LoopDefinition } from '../../engine/LoopLibrary'; 8 | import type { AssetClip } from '../../types/project';"
+  - generic [ref=e7]: at constructor (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:365:19) at TypeScriptParserMixin.raise (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:6599:19) at TypeScriptParserMixin.parseStatementContent (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12871:18) at TypeScriptParserMixin.parseStatementContent (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:9508:18) at TypeScriptParserMixin.parseStatementLike (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12767:17) at TypeScriptParserMixin.parseStatementListItem (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12747:17) at TypeScriptParserMixin.parseBlockOrModuleBlockBody (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13316:61) at TypeScriptParserMixin.parseBlockBody (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13309:10) at TypeScriptParserMixin.parseBlock (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13297:10) at TypeScriptParserMixin.parseFunctionBody (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12100:24) at TypeScriptParserMixin.parseFunctionBodyAndFinish (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12086:10) at TypeScriptParserMixin.parseFunctionBodyAndFinish (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:9172:18) at /Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13445:12 at TypeScriptParserMixin.withSmartMixTopicForbiddingContext (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12403:14) at TypeScriptParserMixin.parseFunction (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13444:10) at TypeScriptParserMixin.parseFunctionStatement (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13125:17) at TypeScriptParserMixin.parseStatementContent (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12791:21) at TypeScriptParserMixin.parseStatementContent (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:9508:18) at TypeScriptParserMixin.parseStatementLike (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12767:17) at TypeScriptParserMixin.parseStatementListItem (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12747:17) at TypeScriptParserMixin.parseExportDeclaration (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13934:17) at TypeScriptParserMixin.parseExportDeclaration (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:9629:85) at TypeScriptParserMixin.maybeParseExportDeclaration (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13893:31) at TypeScriptParserMixin.parseExport (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13812:29) at TypeScriptParserMixin.parseExport (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:9385:20) at TypeScriptParserMixin.parseStatementContent (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12878:27) at TypeScriptParserMixin.parseStatementContent (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:9508:18) at TypeScriptParserMixin.parseStatementLike (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12767:17) at TypeScriptParserMixin.parseModuleItem (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12744:17) at TypeScriptParserMixin.parseBlockOrModuleBlockBody (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13316:36) at TypeScriptParserMixin.parseBlockBody (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:13309:10) at TypeScriptParserMixin.parseProgram (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12622:10) at TypeScriptParserMixin.parseTopLevel (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:12612:25) at TypeScriptParserMixin.parse (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:14488:25) at TypeScriptParserMixin.parse (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:10126:18) at parse (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/parser/lib/index.js:14522:38) at parser (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/core/lib/parser/index.js:41:34) at parser.next (<anonymous>) at normalizeFile (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/core/lib/transformation/normalize-file.js:64:37) at normalizeFile.next (<anonymous>) at run (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/core/lib/transformation/index.js:22:50) at run.next (<anonymous>) at transform (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/core/lib/transform.js:22:33) at transform.next (<anonymous>) at step (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/gensync/index.js:261:32) at /Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/gensync/index.js:273:13 at async.call.result.err.err (/Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/gensync/index.js:223:11) at /Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/gensync/index.js:189:28 at /Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/@babel/core/lib/gensync-utils/async.js:67:7 at /Users/junmingong/.openclaw/workspace/acestep-daw/node_modules/gensync/index.js:113:33
+  - generic [ref=e8]:
+    - text: Click outside, press Esc key, or fix the code to dismiss.
+    - text: You can also disable this overlay by setting
+    - code [ref=e9]: server.hmr.overlay
+    - text: to
+    - code [ref=e10]: "false"
+    - text: in
+    - code [ref=e11]: vite.config.ts
+    - text: .
+```
